refactor: move custom exceptions to separate exceptions.py file
