+++ backend/product/exceptions.py
@@ -1,0 +1,20 @@
+class BrandNotFound(Exception):
+    pass
+
+class ProductNotFound(Exception):
+    pass
+
+class InvalidProductId(Exception):
+    pass
+
+class CategoryNotFound(Exception):
+    pass
+
+class BrandValidationError(Exception):
+    pass
+
+class ProductValidationError(Exception):
+    pass
+
+class CategoryValidationError(Exception):
+    pass
